method syntax getting there
after 100's of rewrites (slight, but not massive, exaggeration...!) I
almost have a workable syntax.  Not 100% sure on either read or action
methods, but they are close - we can use them with the appropriate
syntax and extract various arguments at the right times - but the devil
is in the details so I need to actually attach them into the circuit to
make sure I've got the right things at the right time.